Initial Akeneo API client
This is essentially a copy from the project this spawned from with the
client turned into a class and a config interface provided

I've added some _very_ basic usage instructions to the README and
confirmed that the ESLint and Jest suites run without issue.

The tests need some expansion to test that the calls work properly, but
testing the code against the existing serverless implementation i nthe
parent project will be a good start (ie. nothing should functionally
change other than creating the client in the project.